+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.16.0 - 2026-08-11
 
+### Added
+
+* Create MixedPipeDirectionAnalyzer. [#178](https://github.com/ionide/ionide-analyzers/pull/178)
+
 ### Fixed
 
 * Update FSharp.Analyzers.SDK to `0.37.2`. Checkout the [release notes](https://github.com/ionide/FSharp.Analyzers.SDK/releases/tag/v0.37.2) for details. [#180](https://github.com/ionide/ionide-analyzers/pull/180)